Flatten produced node module

+++ npm-publish.bash
@@ -32,6 +32,18 @@ function main() {
         return 1
     fi
 
+    msg "compiling release version of typescript"
+    if ! tsc -p tsconfig.release.json; then
+        err "typescript release compilation failed"
+        return 1
+    fi
+
+    msg "packaging module"
+    if ! cp release/* .; then
+        err "packaging module failed"
+        return 1
+    fi
+
      # npm honors this
     rm -f .gitignore
